Improve the way of drawing an image
By utilising an extended version of the native drawImage() function to accept the target canvas width and height we can achieve image scaling if needed, such cases would be to scale a small image to fit it's parent canvas width and height fully. Currently the image is being drawn as-is using it's dimensions without considering the parent canvas dimensions.